# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the CharityBay API server.",
        "title": "CharityBay API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/greets": {
            "get": {
                "description": "Show all greets",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "greet"
                ],
                "summary": "Show all greets",
                "operationId": "get-greets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.greetResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a greet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "greet"
                ],
                "summary": "Create a greet",
                "operationId": "create-greet",
                "parameters": [
                    {
                        "description": "Greet object",
                        "name": "greet",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Greet"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.Greet": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Hello, World!"
                }
            }
        },
        "v1.greetResponse": {
            "type": "object",
            "properties": {
                "greets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Greet"
                    }
                }
            }
        },
        "v1.response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "message"
                }
            }
        }
    }
}